+++ src/littlegruz/arpeegee/entities/RPGMeleePlayer.java
@@ -19,7 +19,7 @@ public class RPGMeleePlayer extends RPGPlayer{
       bide = true;
    }
 
-   public RPGMeleePlayer(String name, int level, int gear, int rage, String incomplete, String complete, String party){
+   public RPGMeleePlayer(String name, int level, float gear, int rage, String incomplete, String complete, String party){
       super(name, level, gear, incomplete, complete, party);
       this.rage = rage;
       bideAmt = 0;
