+++ src/littlegruz/arpeegee/entities/RPGRangedPlayer.java
@@ -12,7 +12,7 @@ public class RPGRangedPlayer extends RPGPlayer{
       arrow = true;
    }
 
-   public RPGRangedPlayer(String name, int level, int gear, String incomplete, String complete, String party){
+   public RPGRangedPlayer(String name, int level, float gear, String incomplete, String complete, String party){
       super(name, level, gear, incomplete, complete, party);
       egg = true;
       slowBow = true;
